+++ src/main/java/wooteco/subway/line/domain/LineEntity.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.line.domain;
 
 public class LineEntity {
     private Long id;
