+++ src/main/java/wooteco/subway/section/domain/SectionEntity.java
@@ -1,4 +1,4 @@
-package wooteco.subway.section;
+package wooteco.subway.section.domain;
 
 public class SectionEntity {
     private Long id;
