+++ src/main/java/wooteco/subway/station/service/StationService.java
@@ -1,6 +1,8 @@
-package wooteco.subway.station;
+package wooteco.subway.station.service;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.exception.StationError;
 import wooteco.subway.station.exception.StationException;
