+++ src/main/java/wooteco/subway/station/ui/StationController.java
@@ -1,10 +1,12 @@
-package wooteco.subway.station;
+package wooteco.subway.station.ui;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.service.StationService;
 
 import java.net.URI;
 import java.util.List;
