+++ src/test/java/wooteco/subway/section/SectionsTest.java
@@ -3,9 +3,11 @@ package wooteco.subway.section;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.section.domain.Section;
+import wooteco.subway.section.domain.Sections;
 import wooteco.subway.section.exception.SectionError;
 import wooteco.subway.section.exception.SectionException;
-import wooteco.subway.station.Station;
+import wooteco.subway.station.domain.Station;
 
 import java.util.Arrays;
 import java.util.List;
